Some Mongo
Wow now some pages use MongoDB instead of sql. sql kinda sucked for this
but now with mongo it will be easier for us all to have our own
databases on our own computers because it is so much easier that way to
keep track of what the goddam tables and shit are when with mysql it was
very difficult but with mongo it's easier. So i hope this commit will
help us understand better what the db structure will look like. Thank
you for your time and i hope that this commit finds you well and
prepared to do some coding that will assist us in our ultimate goal of
getting a solid A+ on this website and knocking Gabarro right out of the
park.