Login Page
added some login button code that i copied and pasted from what frank
gave me because he needs me to do that so he can see if his code is
working right. He got like most of it done here i think but there is
more yet to go to get this functionality working all the way so that is
something that i need to commit this for so that he can see all the
changes and modify it really nice so that it works the way that it
should and we can cross that off our checklist of things yet left to do.
That would be great because that is an entire technology that we need to
use so that leaves only three left because this is the social
integration that Gabarro pines for.